requirements.txt
black==21.9b0
click==8.0.3
colorama==0.4.4
haversine==2.5.1
mypy-extensions==0.4.3
numpy==1.21.3
pandas==1.3.4
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
regex==2021.10.23
six==1.16.0
snakeviz==2.1.1
tabulate==0.8.9
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
typing-extensions==3.10.0.2
